backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
httpx==0.26.0
pydantic[email]==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# Data processing
pandas==2.2.0
numpy==1.26.3

# Cardano APIs
blockfrost-python==0.6.0

# Masumi Integration
requests==2.31.0

# AI/LLM (optional)
openai==1.10.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
